+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [3.9.1] - 2020-06-24
+
 ### Fixed
 
 - Fallback on default audio bitrate when absent in mediainfo
@@ -537,7 +539,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Minor fixes and improvements on features and tests
 
-[unreleased]: https://github.com/openfun/marsha/compare/v3.9.0...master
+[unreleased]: https://github.com/openfun/marsha/compare/v3.9.1...master
+[3.9.1]: https://github.com/openfun/marsha/compare/v3.9.0...v3.9.1
 [3.9.0]: https://github.com/openfun/marsha/compare/v3.8.1...v3.9.0
 [3.8.1]: https://github.com/openfun/marsha/compare/v3.8.0...v3.8.1
 [3.8.0]: https://github.com/openfun/marsha/compare/v3.7.1...v3.8.0